requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-pagination==0.12.34
greenlet==3.1.1
h11==0.14.0
idna==3.10
loguru==0.7.3
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
win32_setctime==1.2.0
